+++ cookbooks/groupware/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of groupware.
 
+## 0.3.0:
+
+* recipe/default  - now checks for section presence in box profile's apps list before proceed
+* recipe/hangouts - installs Google Hangouts plugin
+
 ## 0.2.0:
 
 * recipes/skype - delivers first steps documentation concerning skype